README updated

+++ server/models/restaurant_pizza.py
@@ -5,7 +5,7 @@ class RestaurantPizza(db.Model):
 
     id = db.Column (db.Integer, primary_key= True)
 
-    price = db.Column(db.Float, nullable = False) 
+    price = db.Column(db.Integer, nullable = False) 
 
     restaurant_id = db.Column (db.Integer, db.ForeignKey('restaurants.id'))
